static_casts

+++ tasks/seq/Sadikov_I_SparesMatrixMultiplication/perf_tests/main.cpp
@@ -42,7 +42,7 @@ TEST(sadikov_i_sparse_matrix_multiplication_task_seq, test_pipeline_run) {
   auto perf_analyzer = std::make_shared<ppc::core::Perf>(test_task_sequential);
   perf_analyzer->PipelineRun(perf_attr, perf_results);
   ppc::core::Perf::PrintPerfStatistic(perf_results);
-  for (auto i = 0; i < out.size(); ++i) {
+  for (auto i = 0; i < static_cast<int>(out.size()); ++i) {
     EXPECT_NEAR(out[i], checkOut[i], kEpsilon);
   }
 }
@@ -79,7 +79,7 @@ TEST(sadikov_i_sparse_matrix_multiplication_task_seq, test_task_run) {
   auto perf_analyzer = std::make_shared<ppc::core::Perf>(test_task_sequential);
   perf_analyzer->TaskRun(perf_attr, perf_results);
   ppc::core::Perf::PrintPerfStatistic(perf_results);
-  for (auto i = 0; i < out.size(); ++i) {
+  for (auto i = 0; i < static_cast<int>(out.size()); ++i) {
     EXPECT_NEAR(out[i], checkOut[i], kEpsilon);
   }
 }
